requirements.txt
mitmproxy==3.0.4
bitstring==3.1.5
flatten-json==0.1.6
requests==2.18.4
beautifulsoup4==4.6.0
tinycss==0.4
yara-python==3.7.0
yapsy==1.11.223